Hide QR if the attendance is applied and it's not applied by form submissions
